installed: this is a mistake, we need to remove it soon thread.h, libc.h, tos.h: fixes so mk install will work. 9k*: mistake, these must be removed next
R=nix-dev, john
CC=nix-dev
http://codereview.appspot.com/4998043

+++ sys/include/thread.h
@@ -56,7 +56,7 @@ struct Alt {
 };
 
 struct Ref {
-	long	ref;
+	int	ref;
 };
 
 int	alt(Alt alts[]);
